Drop generic webview wrapper
