+++ src/cef/ui/uicefwebview.h
@@ -1,8 +1,11 @@
 #pragma once
 
-#include <framework/ui/uiwebview.h>
+#include <framework/ui/uiwidget.h>
+#include <framework/ui/declarations.h>
 #include <mutex>
 #include <atomic>
+#include <map>
+#include <functional>
 
 #ifdef USE_CEF
 #include "include/cef_browser.h"
@@ -14,11 +17,45 @@
 class SimpleCEFClient;
 #endif
 
-class UICEFWebView : public UIWebView {
+// @bindclass
+class UICEFWebView : public UIWidget {
 public:
     UICEFWebView();
     UICEFWebView(UIWidgetPtr parent);
     virtual ~UICEFWebView();
+
+    // WebView interface
+    void loadUrl(const std::string& url);
+    void loadHtml(const std::string& html, const std::string& baseUrl = "");
+    void loadFile(const std::string& filePath);
+
+    void executeJavaScript(const std::string& script);
+    void setZoomLevel(float zoom);
+    float getZoomLevel();
+
+    void setUserAgent(const std::string& userAgent);
+    std::string getUserAgent();
+
+    void setScrollPosition(const Point& position);
+    Point getScrollPosition();
+
+    void setScrollable(bool scrollable);
+    bool isScrollable();
+
+    void goBack();
+    void goForward();
+    void reload();
+    void stop();
+
+    bool canGoBack();
+    bool canGoForward();
+    bool isLoading();
+
+    void registerJavaScriptCallback(const std::string& name,
+                                    const std::function<void(const std::string&)>& callback,
+                                    int luaRef = -1);
+    void unregisterJavaScriptCallback(const std::string& name);
+    void sendToJavaScript(const std::string& name, const std::string& data);
 
     // CEF-specific methods
     void onPaint(const void* buffer, int width, int height, const CefRenderHandler::RectList& dirtyRects);
@@ -34,11 +71,11 @@ public:
     void setWindowlessFrameRate(int fps);
 
     // Event handlers
-    void onLoadStarted() override;
-    void onLoadFinished(bool success) override;
-    void onUrlChanged(const std::string& url) override;
-    void onTitleChanged(const std::string& title) override;
-    void onJavaScriptCallback(const std::string& name, const std::string& data) override;
+    virtual void onLoadStarted();
+    virtual void onLoadFinished(bool success);
+    virtual void onUrlChanged(const std::string& url);
+    virtual void onTitleChanged(const std::string& title);
+    virtual void onJavaScriptCallback(const std::string& name, const std::string& data);
 
     // Mouse input handlers
     bool onMousePress(const Point& mousePos, Fw::MouseButton button) override;
@@ -57,13 +94,14 @@ public:
     void onGeometryChange(const Rect& oldRect, const Rect& newRect) override;
 
 protected:
-    void createWebView() override;
-    void loadUrlInternal(const std::string& url) override;
-    bool loadHtmlInternal(const std::string& html, const std::string& baseUrl) override;
-    void executeJavaScriptInternal(const std::string& script) override;
     void drawSelf(Fw::DrawPane drawPane) override;
 
 private:
+    void createWebView();
+    void loadUrlInternal(const std::string& url);
+    bool loadHtmlInternal(const std::string& html, const std::string& baseUrl);
+    void executeJavaScriptInternal(const std::string& script);
+
 #ifdef USE_CEF
     friend class SimpleCEFClient;
     CefRefPtr<CefBrowser> m_browser;
@@ -78,9 +116,24 @@ private:
     // Static tracking of all active WebViews (thread-safe)
     static std::vector<UICEFWebView*> s_activeWebViews;
     static std::mutex s_activeWebViewsMutex;
-    
+
     // Instance validity flag (for scheduled events)
     std::atomic<bool> m_isValid;
-    
+
+    // State and JavaScript callbacks
+    std::string m_currentUrl;
+    std::string m_currentTitle;
+    std::string m_userAgent;
+    float m_zoomLevel;
+    Point m_scrollPosition;
+    bool m_scrollable;
+    bool m_loading;
+
+    struct JSCallback {
+        std::function<void(const std::string&)> callback;
+        int luaRef;
+    };
+    std::map<std::string, JSCallback> m_jsCallbacks;
+
 #endif
 };
