+++ src/framework/luafunctions.cpp
@@ -48,7 +48,6 @@
 #include <framework/graphics/particlemanager.h>
 #include <framework/graphics/fontmanager.h>
 #include <framework/ui/ui.h>
-#include <framework/ui/uiwebview.h>
 #ifdef USE_CEF
 #include <cef/ui/uicefwebview.h>
 #endif
