+++ src/framework/ui/ui.h
@@ -32,6 +32,5 @@
 #include "uigridlayout.h"
 #include "uianchorlayout.h"
 #include "uiparticles.h"
-#include "uiwebview.h"
 
 #endif
